Reorganize batch module definitions
