+++ crates/relayer-framework/src/full/batch/impls/message_sender.rs
@@ -1,7 +1,6 @@
 use async_trait::async_trait;
 
 use crate::base::chain::traits::types::HasIbcChainTypes;
-use crate::base::chain::types::aliases::{Event, Message};
 use crate::base::relay::traits::ibc_message_sender::IbcMessageSender;
 use crate::base::relay::traits::target::ChainTarget;
 use crate::base::relay::traits::types::HasRelayTypes;
@@ -10,20 +9,10 @@ use crate::base::runtime::traits::channel::{
 };
 use crate::base::runtime::traits::runtime::HasRuntime;
 use crate::full::batch::traits::channel::HasMessageBatchSender;
+use crate::full::batch::traits::send_messages_from_batch::CanSendIbcMessagesFromBatchWorker;
 use crate::std_prelude::*;
 
 pub struct SendMessagetoBatchWorker;
-
-#[async_trait]
-pub trait CanSendIbcMessagesFromBatchWorker<Target>: HasRelayTypes
-where
-    Target: ChainTarget<Self>,
-{
-    async fn send_messages_from_batch_worker(
-        &self,
-        messages: Vec<Message<Target::TargetChain>>,
-    ) -> Result<Vec<Vec<Event<Target::TargetChain>>>, Self::Error>;
-}
 
 #[async_trait]
 impl<Relay, Target, TargetChain, Runtime> IbcMessageSender<Relay, Target>
