+++ crates/relayer-framework/src/full/batch/impls/spawn.rs
@@ -10,21 +10,13 @@ use crate::base::runtime::traits::sleep::CanSleep;
 use crate::base::runtime::traits::spawn::HasSpawner;
 use crate::base::runtime::traits::time::HasTime;
 use crate::full::batch::traits::channel::HasMessageBatchReceiver;
+use crate::full::batch::traits::send_messages_from_batch::CanSendIbcMessagesFromBatchWorker;
+use crate::full::batch::traits::spawn::CanSpawnBatchMessageWorker;
+use crate::full::batch::types::config::BatchConfig;
+use crate::full::batch::worker::BatchMessageWorker;
 use crate::std_prelude::*;
 
-use super::config::BatchConfig;
-use super::message_sender::CanSendIbcMessagesFromBatchWorker;
-use super::worker::BatchMessageWorker;
-
 pub struct BatchMessageWorkerSpawner<Target>(PhantomData<Target>);
-
-pub trait CanSpawnBatchMessageWorker<Relay, Target>
-where
-    Relay: HasRelayTypes,
-    Target: ChainTarget<Relay>,
-{
-    fn spawn_batch_message_worker(relay: Relay, config: BatchConfig);
-}
 
 impl<Relay, Target, TargetChain, Runtime> CanSpawnBatchMessageWorker<Relay, Target>
     for BatchMessageWorkerSpawner<Target>
