+++ crates/relayer-framework/src/full/batch/mod.rs
@@ -1,6 +1,4 @@
-pub mod config;
-pub mod message_sender;
-pub mod spawn;
+pub mod impls;
 pub mod traits;
 pub mod types;
 pub mod worker;
