+++ crates/relayer-framework/src/full/batch/traits/mod.rs
@@ -1,1 +1,3 @@
 pub mod channel;
+pub mod send_messages_from_batch;
+pub mod spawn;
