+++ crates/relayer-framework/src/full/batch/types/mod.rs
@@ -1,1 +1,2 @@
 pub mod aliases;
+pub mod config;
