+++ crates/relayer-framework/src/full/one_for_all/impls/batch.rs
@@ -8,8 +8,8 @@ use crate::base::relay::traits::ibc_message_sender::IbcMessageSender;
 use crate::base::relay::traits::target::ChainTarget;
 use crate::base::relay::traits::target::{DestinationTarget, SourceTarget};
 use crate::common::one_for_all::types::relay::OfaRelayWrapper;
-use crate::full::batch::message_sender::CanSendIbcMessagesFromBatchWorker;
 use crate::full::batch::traits::channel::{HasMessageBatchReceiver, HasMessageBatchSender};
+use crate::full::batch::traits::send_messages_from_batch::CanSendIbcMessagesFromBatchWorker;
 use crate::full::batch::types::aliases::{MessageBatchReceiver, MessageBatchSender};
 use crate::full::one_for_all::traits::relay::OfaFullRelay;
 use crate::std_prelude::*;
