+++ crates/relayer-framework/src/full/one_for_all/impls/preset.rs
@@ -5,7 +5,7 @@ use crate::base::one_for_all::traits::chain::{OfaChainPreset, OfaIbcChainPreset}
 use crate::base::one_for_all::traits::relay::OfaRelayPreset;
 use crate::base::relay::impls::packet_relayers::general::full_relay::FullRelayer;
 use crate::common::one_for_all::presets::FullPreset;
-use crate::full::batch::message_sender::SendMessagetoBatchWorker;
+use crate::full::batch::impls::message_sender::SendMessagetoBatchWorker;
 use crate::full::filter::impls::filter_relayer::FilterRelayer;
 use crate::full::one_for_all::traits::chain::OfaFullChain;
 use crate::full::one_for_all::traits::relay::OfaFullRelay;
